Fix user profile image bug and goods infomation edit bug

+++ controllers/authControllers/modifiedUser.js
@@ -1,28 +1,32 @@
-require('dotenv').config();
-const jwt = require('jsonwebtoken');
-const fs = require('fs');
-const path = require('path');
-const { User } = require('../../models');
+require("dotenv").config();
+const jwt = require("jsonwebtoken");
+const fs = require("fs");
+const path = require("path");
+const { User } = require("../../models");
 
 module.exports = async (req, res, next) => {
   try {
     const { nick } = req.body;
     const findImage = await User.findOne({
       where: { id: req.user.id },
-      attributes: ['profileImage'],
+      attributes: ["profileImage"],
     });
 
-    if (findImage.profileImage && findImage.profileImage.includes('uploads')) {
+    if (
+      req.file &&
+      findImage.profileImage &&
+      findImage.profileImage.includes("uploads")
+    ) {
       fs.unlink(
-        path.join(__dirname, '../..', findImage.profileImage),
+        path.join(__dirname, "../..", findImage.profileImage),
         (err) => {
           if (err) throw err;
-        },
+        }
       );
     }
     await User.update(
-      { nick, profileImage: req.file.path },
-      { where: { id: req.user.id } },
+      { nick, profileImage: req.file ? req.file.path : findImage.profileImage },
+      { where: { id: req.user.id } }
     );
     const updateInfo = await User.findOne({ where: { id: req.user.id } });
     const token = jwt.sign(
@@ -34,9 +38,9 @@ module.exports = async (req, res, next) => {
         profileImage: updateInfo.profileImage,
       },
       process.env.JWT_SECRET,
-      { expiresIn: '7d' },
+      { expiresIn: "7d" }
     );
-    res.status(200).cookie('access_token', token).json({ access_token: token });
+    res.status(200).cookie("access_token", token).json({ access_token: token });
   } catch (err) {
     next(err);
   }
